pyproject.toml
[project]
name = "openproject-discord-assistant"
version = "0.1.0"
description = "Discord slash command backend for OpenProject work package creation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.1",
  "fastapi>=0.115.7",
  "httpx>=0.28.1",
  "pydantic>=2.10.6",
  "pydantic-settings>=2.8.1",
  "PyNaCl>=1.5.0",
  "sqlalchemy>=2.0.37",
  "psycopg[binary]>=3.2.4",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-mock>=3.14.0",
]

[build-system]
requires = ["setuptools>=75.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
pythonpath = ["."]
